requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.2.0
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
beautifulsoup4==4.12.3
bleach==4.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docker==7.1.0
dropbox==12.0.2
entrypoints==0.4
executing==0.8.3
fastapi==0.103.0
fastjsonschema==2.16.2
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.0
fsspec==2024.6.1
future==1.0.0
gast==0.5.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.63.2
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.48.2
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.24.0
idna==3.7
imagecodecs==2024.6.1
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata==7.1.0
ipykernel==6.28.0
ipython==8.25.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
keras==3.4.1
Keras-Preprocessing==1.1.2
keras-unet==0.1.2
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.5
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.4.0
mlflow==2.14.3
mpmath==1.3.0
multidict==6.0.4
namex==0.0.8
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
optree==0.12.1
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pillow==10.4.0
pip==24.0
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
PyJWT==2.8.0
pyngrok==7.2.0
pyOpenSSL==24.0.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
querystring-parser==1.2.4
referencing==0.30.2
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
rsa==4.7.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==69.5.1
sip==6.7.12
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.2.0
starlette==0.27.0
stone==3.3.1
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.17.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.1.0
terminado==0.17.1
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tinycss2==1.2.1
tokenizers==0.19.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.14.1
yarl==1.9.3
zipp==3.19.2
